requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.0=pyhd8ed1ab_1
arrow-cpp=8.0.0=py310h38b8b19_0
asttokens=2.0.8=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
aws-c-common=0.4.57=ha925a31_1
aws-c-event-stream=0.1.6=hd77b12b_5
aws-checksums=0.1.9=ha925a31_0
aws-sdk-cpp=1.8.185=hd77b12b_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blinker=1.5=pyhd8ed1ab_0
boost-cpp=1.80.0=h9f4b32c_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py310h8d17308_1005
bzip2=1.0.8=h8ffe710_4
c-ares=1.18.1=h8ffe710_0
ca-certificates=2022.4.26=haa95532_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.6.15=py310haa95532_0
cffi=1.15.1=py310h628cb3f_2
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=py310h5588dad_0
colorama=0.4.6=pyhd8ed1ab_0
commonmark=0.9.1=py_0
cryptography=38.0.2=py310h52f42fa_1
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.6.3=py310h8a704f9_0
decorator=5.1.1=pyhd8ed1ab_0
docopt=0.6.2=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.1.1=pyhd8ed1ab_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h546665d_0
future=0.18.2=py310h5588dad_5
gflags=2.2.2=ha925a31_1004
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.29=pyhd8ed1ab_0
glog=0.6.0=h4797de2_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py310h5588dad_0
importlib_resources=5.10.0=pyhd8ed1ab_0
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.16.2=pyh025b116_0
ipython=8.5.0=pyh08f2357_1
ipywidgets=8.0.2=pyhd8ed1ab_1
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h8ffe710_2
jsonschema=4.16.0=pyhd8ed1ab_0
jupyter_client=7.4.4=pyhd8ed1ab_0
jupyter_core=4.11.1=py310h5588dad_0
jupyterlab_widgets=3.0.3=pyhd8ed1ab_0
kiwisolver=1.4.2=py310hd77b12b_0
krb5=1.19.3=h1176d77_0
lcms2=2.12=h2a16943_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libcblas=3.9.0=16_win64_mkl
libcurl=7.85.0=heaf79c2_0
libdeflate=1.14=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=16_win64_mkl
libpng=1.6.38=h19919ed_0
libprotobuf=3.20.1=h12be248_4
libsodium=1.0.18=h8d14728_1
libsqlite=3.39.4=hcfcfb64_0
libssh2=1.10.0=h680486a_3
libthrift=0.15.0=h636ae23_1
libtiff=4.4.0=h8e97e67_4
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libzlib=1.2.13=hcfcfb64_4
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py310he2412df_1
matplotlib-base=3.5.1=py310hd77b12b_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nbformat=5.7.0=pyhd8ed1ab_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numpy=1.23.4=py310h4a8f9c9_0
openjpeg=2.5.0=hc9384bd_1
openssl=1.1.1q=hcfcfb64_1
packaging=21.3=pyhd8ed1ab_0
pandas=1.5.1=py310h1c4a608_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.2.0=py310h52929f7_2
pip=22.3=pyhd8ed1ab_0
pipreqs=0.4.11=pypi_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
prompt-toolkit=3.0.31=pyha770c72_0
protobuf=3.20.1=py310h5588dad_0
psutil=5.9.3=py310h8d17308_0
pthread-stubs=0.4=hcd874cb_1001
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=8.0.0=py310h26aae1b_0
pycparser=2.21=pyhd8ed1ab_0
pydeck=0.8.0b4=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pympler=1.0.1=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.18.1=py310he2412df_1
pysocks=1.7.1=pyh0701188_6
python=3.10.6=h9a09f29_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python-tzdata=2022.5=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.5=pyhd8ed1ab_0
pytz-deprecation-shim=0.1.0.post0=py310h5588dad_2
pywin32=303=py310he2412df_0
pyyaml=6.0=py310he2412df_4
pyzmq=24.0.1=py310hcd737a0_0
re2=2022.04.01=h0e60522_0
requests=2.28.1=pyhd8ed1ab_1
rich=12.6.0=pyhd8ed1ab_0
seaborn=0.12.0=py310haa95532_0
semver=2.13.0=pyh9f0ad1d_0
setuptools=65.5.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smmap=3.0.5=pyh44b312d_0
snappy=1.1.9=h82413e6_1
stack_data=0.5.1=pyhd8ed1ab_0
streamlit=1.13.0=pyhd8ed1ab_1
tbb=2021.6.0=h91493d7_0
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.2=py310he2412df_0
traitlets=5.5.0=pyhd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022e=h191b570_0
tzlocal=4.2=py310h5588dad_1
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.11=pyhd8ed1ab_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.18.2=pyhd3deb0d_0
vc=14.3=h3d8a991_9
vs2015_runtime=14.32.31332=h1d6e394_9
watchdog=2.1.9=py310h5588dad_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=4.0.3=pyhd8ed1ab_0
win_inet_pton=1.1.0=py310h5588dad_5
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
yarg=0.1.9=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.10.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h7755175_4
